Move unread indicator load into SinglePostView
